chore: restructure folder and files
