+++ src/linkedList.spec.ts
@@ -1,4 +1,16 @@
-import { LinkedList } from './linkedList'
+import { LinkedList, Node } from './linkedList'
+
+describe('Node', () => {
+  const instance = new Node('a')
+
+  it('should have properties', () => {
+    expect(instance).toBeInstanceOf(Node)
+    expect(instance).toHaveProperty('value')
+    expect(instance).toHaveProperty('next')
+    expect(instance.value).toBe('a')
+    expect(instance.next).toEqual(null)
+  })
+})
 
 describe('LinkedList', () => {
   let instance: LinkedList
