+++ src/linkedList.ts
@@ -1,4 +1,9 @@
-import { Node } from './node'
+export class Node {
+  next: Node | null
+  constructor(public value: unknown) {
+    this.next = null
+  }
+}
 
 export class LinkedList {
   head: Node | null
