--- src/node.spec.ts
@@ -1,13 +1,0 @@
-import { Node } from './node'
-
-describe('Node', () => {
-  const instance = new Node('a')
-
-  it('should have properties', () => {
-    expect(instance).toBeInstanceOf(Node)
-    expect(instance).toHaveProperty('value')
-    expect(instance).toHaveProperty('next')
-    expect(instance.value).toBe('a')
-    expect(instance.next).toEqual(null)
-  })
-})
