--- src/node.ts
@@ -1,6 +1,0 @@
-export class Node {
-  next: Node | null
-  constructor(public value: unknown) {
-    this.next = null
-  }
-}
